+++ Server/src/main/java/org/elsys/motorcycle_security/controllers/DeviceController.java
@@ -29,8 +29,8 @@ public class DeviceController {
 
     @RequestMapping(value="/device/send/gps-cordinates",method=POST)
     public ResponseEntity sendGpsCordinates(@RequestParam(value="deviceId") String deviceId,
-                                            @RequestParam(value="x", defaultValue="0") long x,
-                                            @RequestParam(value="y", defaultValue="0") long y) {
+                                            @RequestParam(value="x", defaultValue="0") double x,
+                                            @RequestParam(value="y", defaultValue="0") double y) {
         try {
             dataTransmiterHandler.updateGPSCordinates(deviceId, x, y);
         }
